Feature eventos - valida liberação do botão

+++ Backend/plataformaempreendedorismo/src/main/java/com/plataforma/empreendedorismo/plataformaempreendedorismo/repository/EventoRepository.java
@@ -3,5 +3,5 @@ package com.plataforma.empreendedorismo.plataformaempreendedorismo.repository;
 import com.plataforma.empreendedorismo.plataformaempreendedorismo.model.Evento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventoRepository extends JpaRepository<Evento, Integer> {
+public interface EventoRepository extends JpaRepository<Evento, Long> {
 }
